+++ package.json
@@ -48,6 +48,7 @@
     "@delon/form": "^1.0.0-beta.9",
     "@delon/mock": "^1.0.0-beta.9",
     "@delon/theme": "^1.0.0-beta.9",
+    "@delon/util": "^1.0.0-beta.9",
     "@ngx-translate/core": "^9.0.0",
     "@ngx-translate/http-loader": "^2.0.0",
     "@types/ajv": "^1.0.0",
