+++ src/app/routes/dashboard/workplace/workplace.component.ts
@@ -1,7 +1,7 @@
 import { zip } from 'rxjs/observable/zip';
 import { Component, OnInit, OnDestroy } from '@angular/core';
 import { NzMessageService } from 'ng-zorro-antd';
-import { getTimeDistance, yuan, fixedZero } from '@delon/abc';
+import { getTimeDistance, yuan } from '@delon/util';
 import { _HttpClient } from '@delon/theme';
 
 @Component({
