--- src/app/routes/delon/clipboard/clipboard.component.html
@@ -1,10 +1,0 @@
-<div class="content__title">
-  <h1>
-    Clipboard
-    <small>@delon/abc 内置一个简单的
-      <code>copy</code> 方法。</small>
-  </h1>
-</div>
-<nz-card>
-  <button nz-button (click)="onCopy()">Copy</button>
-</nz-card>
