+++ src/app/routes/delon/util/util.component.html
@@ -1,0 +1,54 @@
+<div class="content__title">
+  <h1>
+    工具集
+    <small>@delon/util 日常用法。</small>
+  </h1>
+</div>
+<nz-row [nzGutter]="16">
+  <nz-col [nzSpan]="12">
+    <nz-card nzTitle="字符串类">
+      <nz-card nzType="inner" nzTitle="format">
+        <form nz-form>
+          <nz-form-item>
+            <nz-form-label [nzSm]="8">String</nz-form-label>
+            <nz-form-control [nzSm]="16">
+              <input nz-input [(ngModel)]="format_str" [ngModelOptions]="{standalone: true}">
+            </nz-form-control>
+          </nz-form-item>
+          <nz-form-item>
+            <nz-form-label [nzSm]="8">Object</nz-form-label>
+            <nz-form-control [nzSm]="16">
+              <input nz-input [(ngModel)]="format_obj" [ngModelOptions]="{standalone: true}">
+            </nz-form-control>
+          </nz-form-item>
+          <nz-form-item>
+            <nz-form-control [nzSpan]="16" [nzOffset]="8">
+              <button (click)="onFormat()" nz-button>Run</button>
+            </nz-form-control>
+          </nz-form-item>
+          <nz-form-item>
+            <nz-form-label [nzSm]="8">Result</nz-form-label>
+            <nz-form-control [nzSm]="16">
+              {{format_res}}
+            </nz-form-control>
+          </nz-form-item>
+        </form>
+      </nz-card>
+      <nz-card nzType="inner" nzTitle="yuan">
+        <nz-row [nzGutter]="16">
+          <nz-col [nzSpan]="12">
+            <input type="number" nz-input [(ngModel)]="yuan_str" (ngModelChange)="onYuan($event)">
+          </nz-col>
+          <nz-col [nzSpan]="12">
+            <div [innerHTML]="yuan_res"></div>
+          </nz-col>
+        </nz-row>
+      </nz-card>
+    </nz-card>
+  </nz-col>
+  <nz-col [nzSpan]="12">
+    <nz-card nzTitle="其它类">
+      <button nz-button (click)="onCopy()">Copy</button>
+    </nz-card>
+  </nz-col>
+</nz-row>
