+++ src/app/routes/style/colors/colors.component.ts
@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
 import { ColorService } from '../color.service';
-import { copy } from '@delon/abc';
+import { copy } from '@delon/util';
 import { NzMessageService } from 'ng-zorro-antd';
 
 @Component({
