+++ src/assets/i18n/en.json
@@ -28,7 +28,7 @@
   "delon": "@Delon",
   "dynamic-form": "Dynamic Form",
   "simple-table": "Simple table",
-  "clipboard": "Clipboard",
+  "util": "Util",
   "print": "Print",
   "guard": "Route Guard",
   "cache": "Cache",
